Refactor phase 1: extract Kernel hierarchy into sigway/kernels.py
Move the kernel functions (get_u/get_v/polynomial, I_sq_RD/I_sq_RD_uv/I_sq_MD,
the sici helpers, and the eMD I_sq_IRD_LV/_res + grads) verbatim out of
omega_gw_jax.py into a new sigway/kernels.py, kept as module-level jit-able
cores. Add a Kernel base + RadiationKernel + InstantEMDKernel that wrap those
cores, declare integration structure (k_dependent, param_names, resonant_t) and
own the normalisation (norm set at init, per-kernel default, NORM_PRESETS,
unknown-preset error). The factor-of-2 is folded into kernel.norm.

No behaviour change: OmegaGWjax/OmegaGWms still use the (now imported) functions
and keep their string/shim API; the new classes are not wired in yet (Phase 3).
Repoint internals-facing test imports to sigway.kernels and add tests for the
new classes. flake8/black clean; 51 passed.
