+++ tests/test_omega_gw_jax.py
@@ -5,6 +5,7 @@ import numpy as np
 
 # Local
 from sigway import omega_gw_jax as og
+from sigway.kernels import I_sq_RD_uv
 
 # Load test data
 test_data = np.load(
@@ -82,7 +83,7 @@ class TestUnits(unittest.TestCase):
         # compute I_sq_RD from some values of s and t
         I_sq_RD = og.I_sq_RD(tt[:, None], s[None, :], k=1.0)
 
-        I_sq_RD_2 = og.I_sq_RD_uv(tt[:, None], s[None, :], k=1.0)
+        I_sq_RD_2 = I_sq_RD_uv(tt[:, None], s[None, :], k=1.0)
 
         # check that the output is correct
         self.assertAlmostEqual(np.sum(I_sq_RD / I_sq_RD_2 - 1), 0.0, places=7)
